% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_cov_mat}
\alias{est_cov_mat}
\title{covariance matrix of 10 SNPs}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 10 columns.
}
\source{
{Data is simulated using Hapgen2 and a multi-SNP logistic regression
  model is fitted. This is the covariance matrix of the 10 estimated log-odds
  ratios obtained from the regression output.}
}
\usage{
est_cov_mat
}
\description{
A 10 by 10 covariance matrix containing the pairwise covariances of the
estimates of 10 SNP effect sizes obtained from a multi-SNP logistic
regression model.
}
\examples{
data(est_cov_mat)
}
\keyword{datasets}
