% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{est_log_ors}
\alias{est_log_ors}
\title{Effect size estimates of 10 SNPs}
\format{
An object of class \code{numeric} of length 10.
}
\source{
{Data is simulated using Hapgen2 and a multi-SNP logistic regression
  model is fitted. These are the estimated log-odds ratios from the
  regression output.}
}
\usage{
est_log_ors
}
\description{
A vector of length 10 containing the estimates of 10 SNP effect sizes
obtained from a multi-SNP logistic regression model.
}
\examples{
data(est_log_ors)
}
\keyword{datasets}
