% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaussian_multivar.R
\name{gauss_mv}
\alias{gauss_mv}
\title{Calculate PPI using the Gaussian prior}
\usage{
gauss_mv(num_snps, beta_hats, V_inv, maxk, w, prior_probs, allow_no_causal)
}
\arguments{
\item{num_snps}{The number of SNPs in your dataset.}

\item{beta_hats}{A vector of estimates of the SNP effect sizes (log-odds
ratios).}

\item{V_inv}{The inverse of the covariance matrix (of dim \code{num_snps} by
\code{num_snps}) of the estimated SNP effect sizes.}

\item{maxk}{The maximum number of causal SNPs allowed in the model.}

\item{w}{The Gaussian effect size prior variance.}

\item{prior_probs}{A vector of prior probabilities for the number of causal
SNPs. For example if you want to allow no more than two causal SNPs with
either one and two causal SNPs equally likely a priori then use
\code{c(0.5, 0.5)}. You can allow for the possibility that there are no
causal SNPs but the flag \code{allow_no_causal} must be \code{TRUE}. For
example to specify that the prior probability of no causal SNPs is 0.5 and
that the prior probability of either one or two causal SNPs is 0.25 use
\code{c(0.5, 0.25, 0.25)}.}

\item{allow_no_causal}{Are you allowing for the possibility that there are no
causal SNPs? I.e. is the first element of \code{prior_probs} the
probability that there are no causal SNPs a priori?}
}
\value{
A list consisting of (in order) the posterior inclusion probability
  by SNP, the posterior probability of each model, the prior_probabilities
  used (these are the same as those specified in the function arguments).
  Additionally if \code{allow_no_causal} is \code{TRUE} then the posterior
  probability of the null model (no causal SNPs) is also returned as the
  fourth list element.
}
\description{
Calculate PPI using the Gaussian prior
}
\examples{
out <- gauss_mv(num_snps = 10, beta_hats = est_log_ors,
V_inv = solve(est_cov_mat), maxk = 2, w = 0.04,
prior_probs = c(0.5, 0.5), allow_no_causal = FALSE)
out[[1]] # PPIs of the 10 SNPs
out[[2]] # Posterior probabilities of the 1-SNP and 2-SNP models
out[[3]] # prior probabilities of the number of causal SNPs
}
